# backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TherapEase API",
        "version": "1.0.0",
        "description": "Swagger docs for the TherapEase psychotherapy platform"
    },
    "servers": [
        {
            "url": "http://localhost:5000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "summary": "Register a new user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "firstName",
                                    "lastName",
                                    "phone",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "firstName": {
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registration created or pending approval"
                    },
                    "400": {
                        "description": "Missing or invalid data"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "summary": "Log in a user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged in (returns user+token)"
                    },
                    "400": {
                        "description": "Invalid credentials"
                    },
                    "403": {
                        "description": "Client not yet approved"
                    }
                }
            }
        },
...
    }
}